Add more intuitive deleting of dependents with Apply
Currently, in order to delete dependent objects with Apply, one must use
`Apply.WithOwner.ApplyObject()`, which does not convey that any
deletions are happening.

After this change, Apply will now have a `DeleteDependents` method that
will wrap the above call, but is much more intuitive to a user of the
interface.